renaming breit-wigner classes
