+++ examples/bat_gen/models/d3pi.h
@@ -30,7 +30,6 @@
 #include <PDL.h>
 #include <PoleMass.h>
 #include <QuantumNumbers.h>
-#include <RelativisticBreitWigner.h>
 #include <SpinAmplitudeCache.h>
 
 #include <BAT/BCGaussianPrior.h>
@@ -59,12 +58,12 @@ inline unique_ptr<Model> d3pi(unique_ptr<Model> M)
     auto D = DecayingParticle::create(T["D+"], radialSize);
 
     // rho
-    auto rho = DecayingParticle::create(T[113], radialSize, make_shared<RelativisticBreitWigner>(775.49e-3, 149.4e-3));
+    auto rho = DecayingParticle::create(T[113], radialSize, make_shared<BreitWigner>(775.49e-3, 149.4e-3));
     rho->addStrongDecay(piPlus, piMinus);
     D->addWeakDecay(rho, piPlus);
 
     // f_2(1270)
-    auto f_2 = DecayingParticle::create(T[225], radialSize, make_shared<RelativisticBreitWigner>(T[225]));
+    auto f_2 = DecayingParticle::create(T[225], radialSize, make_shared<BreitWigner>(T[225]));
     f_2->addStrongDecay(piPlus, piMinus);
     D->addWeakDecay(f_2, piPlus);
 
@@ -77,12 +76,12 @@ inline unique_ptr<Model> d3pi(unique_ptr<Model> M)
     D->addWeakDecay(f_0_980, piPlus);
 
     // f_0(1370)
-    auto f_0_1370 = DecayingParticle::create("f_0(1370)", T["f_0"].quantumNumbers(), radialSize, make_unique<RelativisticBreitWigner>(1.350, 0.265));
+    auto f_0_1370 = DecayingParticle::create("f_0(1370)", T["f_0"].quantumNumbers(), radialSize, make_unique<BreitWigner>(1.350, 0.265));
     f_0_1370->addStrongDecay(piPlus, piMinus);
     D->addWeakDecay(f_0_1370, piPlus);
 
     // f_0(1500)
-    auto f_0_1500 = DecayingParticle::create(T["f_0(1500)"], radialSize, make_unique<RelativisticBreitWigner>(T["f_0(1500)"]));
+    auto f_0_1500 = DecayingParticle::create(T["f_0(1500)"], radialSize, make_unique<BreitWigner>(T["f_0(1500)"]));
     f_0_1500->addStrongDecay(piPlus, piMinus);
     D->addWeakDecay(f_0_1500, piPlus);
